refactored backend API logic
